skills/git-analyzer-skill/requirements.txt
GitPython>=3.1.0
pygit2>=1.10.0
click>=8.0.0
